+++ cognite/experimental/_api/vision.py
@@ -8,6 +8,7 @@ from cognite.experimental.data_classes.vision import (
     DetectAssetsInFilesJob,
     EitherFileId,
     Feature,
+    FeatureParameters,
     InternalId,
     VisionExtractJob,
 )
@@ -76,8 +77,9 @@ class VisionAPI(ContextAPI):
         features: Union[Feature, List[Feature]],
         file_ids: Optional[List[int]] = None,
         file_external_ids: Optional[List[str]] = None,
+        parameters: Optional[FeatureParameters] = None,
     ) -> VisionExtractJob:
-        """Extract features from image files.
+        """Start an asynchronous job to extract features from image files.
 
         Args:
             features (Union[Feature, List[Feature]]): The feature(s) to extract from the provided image files.
@@ -113,11 +115,12 @@ class VisionAPI(ContextAPI):
             status_path="/extract/",
             items=self._process_file_ids(file_ids, file_external_ids),
             features=features,
+            parameters=parameters.dump(camel_case=True) if parameters is not None else None,
             job_cls=VisionExtractJob,
         )
 
     def get_extract_job(self, job_id: InternalId) -> VisionExtractJob:
-        """Retrieve an extraction job.
+        """Retrieve an existing extract job by ID.
 
         Args:
             job_id (InternalId): ID of an existing feature extraction job.