+++ tests/tests_unit/test_api/good_absolute_import/shared/util.py
@@ -1,2 +1,0 @@
-def shared_func():
-    return 42