+++ tests/tests_unit/test_api/relative_imports/bad_relative_import.py
@@ -1,5 +1,0 @@
-from .util import local_func
-
-
-def handle():
-    return local_func()